requirements-travis.txt
awscli
boto3
gitpython
pandas
sagemaker
